bisection workaround (preliminary)
In the previous versions, L-BFGS inversions tended to get stuk in the 'bisection loop' in optlib_wolfe. This is (presumably) because the calculated gradient is incorrect, whereupon the inversion can't find a L-BFGS update which satisfies both the Armijo-Goldstein and Wolfe conditions. In this bisection loop, it just keeps calculating models and misfits for new step lengths, the difference between which becomes tinier and tinier. Meanwhile, it generates huge amounts of data which are written to disk.

Now, the bisection loop is limited to 5 trials. If a model is then still not found, either
(i) the Armijo step is taken, if this step is larger than 0.01 (otherwise Armijo is already a bit dubious), or
(ii) if the Armijo step is tiny as well, the inversion is simply aborted.

Also some minor changes:
- when running forward, the percentage completed will disappear once the wave propagation is completed.
- plot_model_diff now takes model numbers instead of only already created models.

TODO: Ultimately, in the case (ii) one should go back to the steepest descent step (calculating a suitable step length), and see if that gives a decrease in misfit at all. But this remains a TODO.