patch to fix_velocities
there was a bug in the inversion if fix_velocities was turned on and the starting model and background model were not the same. map_m_to_parameters would calculate mu and lambda from the background vp and vs values, which would mean that vp and vs were fixed to the BACKGROUND values, instead of to the values from the STARTING MODEL.

I changed map_m_to_parameters so that instead of vp_bg and vs_bg, vp and vs from the starting model are used.

Other than that, some cosmetics:
- inversion result plot now fits in screen
- in the inversion_routine, Model_start is now used and Model(1) is only defined at new_iteration
- the L-BFGS inversion result shows the angle between kernels right up until the current iteration