added kernel patch to ignore bottom rows
Before this patch, the wave propagation code resulted in focusing of the kernel at the bottom of the domain as the iterations progressed. This was most probably a result of the finite difference scheme, which cannot properly deal with the rigid boundary conditions imposed in the current mantle-like models at the bottom. (At the top, where the boundary is also rigid, this problem is mitigated with lots of coverage (sources and receivers near the top of the domain. At the sides, absorbing boundaries are implemented).

Christian suggested to ignore the bottom 5 rows of the kernel in each update (i.e. set them to zero), because this is the width of the finite difference stencil. I included a switch in the input to 'zero_bottom_rows', and indeed the problem is (almost) gone.

Additionally, some cosmetics and more general code improvements:
- in optlib_wolfe, the Armijo attempt cutoff is decreased from sig < 0.001*sig0 to sig < 0.01*sig0 (because it wasn't leading anywhere anyway)
- the smoothing and reparametrising and making-relative-kernels operations were all moved from the eval_ to the map_ functions (I find this cleaner)
- a new function was made: calculate_unrelative_kernels (because I needed it in map_gradm_to_gradparameters)
- plot_inversion_output was slightly changed for better visuals

TODO:
I probably should change the bottom 5 rows to bottom 6 rows. This is (I think) because the actual bottom row is fixed, and therefore the FD scheme has no influence there. The 5 rows _above_ that, on the other hand, will be influenced by the boundary condition, so I should set all of those to zero."