Source-time-function improvement
Previously, a different dt would result in a different source-time-function amplitude and thus different seismograms. This had to do with two things:
- src_info(ii).f_max was set to 1 Hz, while simultaneously dt was also 1 second.
- the source-time-function was a filtered delta pulse which was not normalised by dt, and therefore the integral over time was not equal to 1.

The code-wise solution was to normalise the stf by its maximum amplitude after filtering it for the first time (in make_source_time_function) (an alternative, I think, would be to divide the initial delta amplitude by dt). Furthermore, I decreased src_info(ii).f_max to 0.1 Hz, which is well below the sampling rate.

In general, care should be taken with src_info(ii).f_max.!!

Additionally, I added source_amplitude to input_parameters, so that one also has influence on the magnitude of the source.