some smaller debugs & plotting things
added an if-switch to eval_obj&grad and eval_grad that only those kernels are saved to currentKernels that actually exist. This bug was slumbering, because I was using seis+grav all the time in my inversions, but would have become a problem as soon as I used only seis or only grav.
in prepare_obs, I moved the plotting of seismograms to the end of each frequency. (could even move it to the end of prepare_obs altogether)
added 'delta' as possible source-time-function (unfiltered)
changed the saving of video in run_fwd and run_adj to 'Motion JPEG AVI' (from 'Uncompressed AVI') - saves massive amounts of space.

Furthermore, I played around with figure plotting:
- see if I can set visibility of figures to off (didn't work for some reason, haven't looked into further)
- changed '-r400' to '-r0' to see if this solves the saving-only-half-the-figure issue when using matlab in nodisplay mode (e.g. remotely) -- but it didn't.