updated saving .mat-files
if a matfile turns out to be bigger than 2 Gb, it will now save as a -v7.3 file instead of -v6. (always as -v7.3 is not an option because it is SUPER slow)

NOTE: Right now, I've implemented this as a 'whos' command where the total size of variables is added up, and I've taken a rather arbitrary limit of 2.1472e9  bytes -- which simply SEEMED to be the limit where it would stop saving as -v6.