added options.grav_step_length
In the previous cases, if a 'Warning: Bad angle of search direction. Use steepeste descent instead' was encountered in optlib_lbfgs, optlib_wolfe wasn't then able to find a good step length for the steepest descent step: a step of 1 was too big, and after several halving operations it would reach the Armijo validity limit of 0.01, after which it quit.

A makeshift solution was to supply the initial stepInit (from outside the inversion toolbox) as 'options.grad_step_length' to the Inversion Toolbox, which is now the step used whenever a steepest descent step is encountered. This seems to work very well, so far, although the supplied step length in these cases is probably smaller than necessary.

TODO: this coul be further optimised by letting optlib_wolfe search for bigger steps in the case where the grad_step_length was used.