Added resume_iterations_atfreq
It was such a pain to restart iterations at a certain frequency if something had crashed, that I decided to make a special script for this. The starting frequency is hardcoded in this script, so it needs to be changed every time you want to youse it, but at least it's much easier.